Remove DEFAULT_OUT
This was kinda hacky and wouldn't work very well when arch-chrooting as a user
or arch-chrooting as root and then switching to a user since /dev/stdout would
be owned by root causing permission issues.
To keep it simple, it is planned to not redirect command output to /dev/null in
most circumstances and to use -q/--quiet/-s/--silent flags instead.