Fix build errors about invalid cross-device links
So here's what was happening. The host system's pacman hooks would be executed
when pacstrap-ing the iso. This would cause refind-install to be executed by
the host's refind hook. The refind-install script would not find the boot
partition while its chrooted in the ISO and would automatically mount the host
system's boot partition into the boot directory in the chroot. After the chroot
when the build script is copying files from airootfs and making hardlinks to
the work directory, it would try to make a hard link from the now-mounted boot
partition to the work directory which exists on a different paritition (/var)
causing the invalid cross-device link error.