Split feature_common into header and footer
Keeping all the common code in one file included at the bottom caused problems
since some functions/statements needed to be sourced before the feature rather
than later (i.e. VERBOSITY_FLAGS) would not be usable in features unless they
were in functions since these flags were sourced at the end. Splitting up
this file into a header and footer allows some code to be sourced before the
feature and some after for flexibility.

This fixes the sudo aliasing issue which never actually worked since the alias
would have to be set before the rest of the feature code. Sourcing this alias
at the bottom of the feature never actually worked.

Moving the sudo alias hack to the header revealed another issue with the alias,
aliasing it to nothing caused the install command to invoke the install function
causing an infinite recursive loop. Aliasing sudo to 'command' fixes this.

Also err_trap was moved to the top so all errors will trigger the function.
Having this function at the bottom caused the interpreter to not be able to call
the function if the error took place before the function's definition.