Install recovery mode on a separate partition
Install recovery mode on separate partition for extra safety and keep the
partition unmounted. This way if anything happens to the main boot partition
or anything else on the system, the recovery partition is much more likely to
remain untouched.

In addition, this also allows modifying the system's boot partition. Before,
when the recovery image was on the system's boot partition, the system's boot
partition could not be mounted or modified when in recovery mode. Now with
recovery mode on a separate partition, the entire system is modifiable and
mountable.