Fix mkinitcpio kernel image not found error
For some reason pacstrap does not install the latest version of the linux
kernel during the build. Also /usr/lib/modules/*/vmlinuz expands to the host's
path which does not match the iso path. The reason this was changed from the
original script was the build was not working when the bug from commit
0d4038dab0d51c40c9712d4b9d6904750edbf270 was causing the host's boot partition
to be mounted on the archiso boot partition. After resolving that issue, the
original code works fine.