Set TPM flag onces at top of luks-encryption-tpm
Before when feature_header.sh and feature_footer.sh were combined, any code
outside of the functions would not be executed since the feature.sh file would
source the feature_common.sh file which meant any code executed in the common
file would execute before any code in the feature.sh file.

Now that feature_footer.sh is sourced after all of feature.sh, it is possible
to execute code in feature.sh outside of functions.

Originally this code was throwing errors since readonly variables could only be
set once.  The way it was, was causing the variable to be assigned multiple
times.