Use blkid to check encryption in user-install.sh
A template crypttab file is created in /etc which means the existance of a
crypttab file is not a reliable indicator of encryption being setup. Instead,
look for a 'crypt' in blkid.

Also remove the sudoers file at the end, so sudo can be used to retreive the
contents of the crypttab file during the review configuration stage.