Load card name from conf.sh for wpa-supplicant
Try to load the card name from conf.sh when installing the wpa-supplicant
feature to remove any need for manual input. This will help automating the
install.

Also use find to check if wpa_supplicant-*.conf exists since it is more
reliable.

Remove check_conf function, otherwise a non-zero return of that function would
cause the script to exit. It's okay if either of the configuration files are not
properly set up, since the script will simply make the process more interactive.

This fixes issue #8